requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangGraph & LangChain (actualizados y compatibles)
langchain>=0.2,<0.4
langgraph>=0.2.20,<0.3
langchain-google-vertexai>=0.1,<0.2

# Google Cloud
google-cloud-bigquery==3.15.0
google-cloud-aiplatform==1.40.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0

# Utils
httpx==0.26.0
python-multipart==0.0.6

# Development (opcional)
pytest==7.4.4
black==23.12.1
